itemDetail refactor
