+++ frontend/App/Components/Header/HeaderRight.tsx
@@ -1,10 +1,5 @@
 import React, { Component } from 'react';
 import { View } from 'react-native';
-import { Icon } from 'react-native-elements';
-import { Colors } from '../../Themes/';
-
-import * as NavigationConfig from '../../Navigation/NavigationConfig';
-import { reverse } from 'dns';
 
 export default class HeaderRight extends Component {
     constructor(props) {
