+++ frontend/App/Containers/Styles/ItemDetailScreenStyle.ts
@@ -1,5 +1,5 @@
 import { StyleSheet } from 'react-native';
-import { ApplicationStyles, Colors } from '../../Themes/';
+import { ApplicationStyles, Colors } from '../../Themes';
 
 export default StyleSheet.create({
     ...ApplicationStyles.screen,
@@ -51,59 +51,6 @@ export default StyleSheet.create({
         fontSize: 10,
         marginLeft: 10,
         fontStyle: 'italic'
-    },
-    castHeader: {
-        marginTop: 10,
-        fontSize: 16
-    },
-    castContainer: {
-        flex: 1, 
-        marginHorizontal: 15
-    },
-    castName: {
-        width: 75, 
-        textAlign: 'center', 
-        fontWeight: 'bold'
-    },
-    castCharacter: {
-        width: 75, 
-        textAlign: 'center', 
-        fontSize: 10, 
-        fontStyle: 'italic'
-    },
-    avatarContainer: {
-        flexDirection: 'row', 
-        marginTop: 5
-    },
-    genreContainer: {
-        flex: 1, 
-        flexDirection: 'row', 
-        flexWrap: 'wrap', 
-        marginLeft: 13, 
-        marginRight: 15
-    },
-    seasonsContainer: {
-        flex: 1, 
-        marginHorizontal: 15
-    },
-    seasonsHeader: {
-        marginTop: 10,
-        fontSize: 16
-    },
-    seasonsName: {
-        width:75, 
-        textAlign: 'center', 
-        fontWeight: 'bold'
-    },
-    buttonsContainer: {
-        flex: 1, 
-        flexDirection: 'row', 
-        marginVertical: 10
-    },
-    divider: {
-        backgroundColor: Colors.divider, 
-        height: 1, 
-        marginTop: 10
     },
     descriptionContainer: {
         marginTop: 60,
